Move CallableTemplateMixin into _meta_mixins
* This means _meta_mixins is self-contained from other Hera code (i.e. mixins)
