+++ src/hera/workflows/container_set.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from typing import Any, List, Optional, Union
 
+from hera.workflows._meta_mixins import CallableTemplateMixin
 from hera.workflows._mixins import (
     ArgumentsMixin,
-    CallableTemplateMixin,
     ContainerMixin,
     ContextMixin,
     EnvIOMixin,
