+++ src/hera/workflows/suspend.py
@@ -9,7 +9,8 @@ See https://argoproj.github.io/argo-workflows/intermediate-inputs/ for more on i
 
 from typing import List, Optional, Union
 
-from hera.workflows._mixins import ArgumentsMixin, CallableTemplateMixin, TemplateMixin
+from hera.workflows._meta_mixins import CallableTemplateMixin
+from hera.workflows._mixins import ArgumentsMixin, TemplateMixin
 from hera.workflows.models import (
     Inputs,
     Outputs,
